GS-9994 Fixed type error in configurer
svn path=/xap/trunk/openspaces/; revision=113909


Former-commit-id: 96d273b7ca23e7f5a78b77c6bf9ad4e37ec7a89c

+++ src/main/src/org/openspaces/admin/pu/statistics/PercentileTimeWindowStatisticsConfigurer.java
@@ -25,7 +25,7 @@ public class PercentileTimeWindowStatisticsConfigurer {
         return this;
     }
     
-    public PercentileTimeWindowStatisticsConfigurer percentile(int percentile) {
+    public PercentileTimeWindowStatisticsConfigurer percentile(double percentile) {
         config.setPercentile(percentile);
         return this;
     }
@@ -34,5 +34,4 @@ public class PercentileTimeWindowStatisticsConfigurer {
         config.validate();       
         return config;
     }
-    
 }
